requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
altair==4.0
    # via streamlit
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
backports-zoneinfo==0.2.1
    # via tzlocal
blinker==1.7.0
    # via streamlit
cachetools==5.3.2
    # via streamlit
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   nltk
    #   streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
dnspython==2.5.0
    # via pymongo
evidently==0.2.6
    # via -r requirements.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.41
    # via streamlit
idna==3.6
    # via requests
importlib-metadata==7.0.1
    # via streamlit
importlib-resources==6.1.1
    # via
    #   jsonschema
    #   jsonschema-specifications
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
joblib==1.3.2
    # via
    #   nltk
    #   scikit-learn
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.4
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
nltk==3.8.1
    # via evidently
numpy==1.24.2
    # via
    #   -r requirements.in
    #   altair
    #   evidently
    #   pandas
    #   patsy
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   statsmodels
    #   streamlit
packaging==23.2
    # via
    #   altair
    #   plotly
    #   statsmodels
    #   streamlit
pandas==1.5.3
    # via
    #   -r requirements.in
    #   altair
    #   evidently
    #   statsmodels
    #   streamlit
patsy==0.5.6
    # via statsmodels
pillow==10.2.0
    # via
    #   -r requirements.in
    #   streamlit
pkgutil-resolve-name==1.3.10
    # via jsonschema
plotly==5.18.0
    # via evidently
protobuf==3.20.3
    # via streamlit
pyarrow==11.0.0
    # via
    #   -r requirements.in
    #   streamlit
pydeck==0.8.1b0
    # via streamlit
pygments==2.17.2
    # via rich
pymongo==4.6.1
    # via -r requirements.in
pympler==1.0.1
    # via streamlit
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==5.4.1
    # via evidently
referencing==0.32.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via nltk
requests==2.31.0
    # via
    #   evidently
    #   streamlit
rich==13.7.0
    # via streamlit
rpds-py==0.17.1
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.3.2
    # via evidently
scipy==1.10.1
    # via
    #   evidently
    #   scikit-learn
    #   statsmodels
semver==3.0.2
    # via streamlit
six==1.16.0
    # via
    #   patsy
    #   python-dateutil
smmap==5.0.1
    # via gitdb
statsmodels==0.14.1
    # via evidently
streamlit==1.19.0
    # via -r requirements.in
tenacity==8.2.3
    # via plotly
threadpoolctl==3.2.0
    # via scikit-learn
toml==0.10.2
    # via streamlit
toolz==0.12.0
    # via altair
tornado==6.4
    # via streamlit
tqdm==4.66.1
    # via nltk
types-python-dateutil==2.8.19.20240106
    # via -r requirements.in
typing-extensions==4.9.0
    # via
    #   -r requirements.in
    #   altair
    #   rich
    #   streamlit
tzdata==2023.4
    # via tzlocal
tzlocal==5.2
    # via streamlit
urllib3==2.1.0
    # via requests
validators==0.22.0
    # via streamlit
watchdog==3.0.0
    # via streamlit
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources
